refactor: llamalend table active filters
